+++ src/Entity/Submission.php
@@ -233,8 +233,6 @@ class Submission
     public const SUBMISSION_STATE_DRAFT = 0b00000001;
     // leave empty for potential state between draft and submitted (sneak preview?)
     public const SUBMISSION_STATE_SUBMITTED = 0b00000100;
-    // leave empty for potential state between submission and accepted (review?)
-    public const SUBMISSION_STATE_ACCEPTED = 0b00010000;
 
     #[ORM\Id]
     #[ORM\Column(type: 'string', length: 50, nullable: false)]
@@ -387,12 +385,6 @@ class Submission
     public function isDraft(): bool
     {
         return $this->submissionState === self::SUBMISSION_STATE_DRAFT;
-    }
-
-    #[Ignore]
-    public function isAccepted(): bool
-    {
-        return $this->submissionState === self::SUBMISSION_STATE_ACCEPTED;
     }
 
     public function getSubmittedFiles(): Collection